Modify setting structures.
