--- app/src/main/res/layout/activity_settings.xml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-    <data>
-        <variable name="activity" type="jp.toastkid.calendar.settings.SettingsActivity" />
-    </data>
-    <FrameLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        >
-        <LinearLayout
-            android:orientation="vertical"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/soft_background"
-            >
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/settings_toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="56dp"
-                android:minHeight="?android:actionBarSize"
-                />
-            <FrameLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                >
-                <ImageView
-                    android:id="@+id/setting_background"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:scaleType="centerCrop"
-                    android:adjustViewBounds="true"
-                    />
-                <LinearLayout
-                    android:orientation="vertical"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:background="@color/filter_white_bb"
-                    >
-                    <LinearLayout
-                        android:id="@+id/settings_color"
-                        android:orientation="vertical"
-                        android:layout_width="match_parent"
-                        android:layout_height="@dimen/settings_item_height"
-                        android:paddingLeft="@dimen/settings_item_left_margin"
-                        android:background="@drawable/ripple_transparent"
-                        android:onClick="@{activity.color}"
-                        >
-                        <TextView
-                            android:id="@+id/settings_color_text"
-                            android:layout_width="match_parent"
-                            android:layout_height="match_parent"
-                            android:gravity="center_vertical"
-                            android:text="@string/title_settings_color"
-                            android:textSize="16sp"
-                            android:drawableLeft="@drawable/ic_palette"
-                            />
-                    </LinearLayout>
-
-                    <include layout="@layout/primary_dark_divider" />
-
-                    <FrameLayout
-                        android:id="@+id/settings_enable_suggest"
-                        android:layout_width="match_parent"
-                        android:layout_height="@dimen/settings_item_height"
-                        android:background="@drawable/ripple_transparent"
-                        android:onClick="@{activity.switchSuggest}"
-                        >
-
-                        <TextView
-                            android:id="@+id/settings_enable_suggest_text"
-                            android:layout_width="wrap_content"
-                            android:layout_height="match_parent"
-                            android:layout_marginLeft="@dimen/settings_item_left_margin"
-                            android:drawableLeft="@drawable/ic_assist"
-                            android:gravity="center_vertical"
-                            android:text="@string/title_enable_suggest"
-                            android:textSize="16sp"
-                            />
-
-                        <CheckBox
-                            android:id="@+id/settings_enable_suggest_check"
-                            android:layout_width="wrap_content"
-                            android:layout_height="match_parent"
-                            android:layout_gravity="right"
-                            android:layout_marginRight="@dimen/settings_item_left_margin"
-                            android:clickable="false"
-                            android:gravity="center_vertical"
-                            />
-
-                    </FrameLayout>
-
-                    <include layout="@layout/primary_dark_divider" />
-
-                    <TextView
-                        android:id="@+id/settings_clear"
-                        android:layout_width="match_parent"
-                        android:layout_height="@dimen/settings_item_height"
-                        android:drawableLeft="@drawable/ic_close_black"
-                        android:gravity="center_vertical"
-                        android:paddingLeft="@dimen/settings_item_left_margin"
-                        android:text="@string/title_clear"
-                        android:textSize="16sp"
-                        android:background="@drawable/ripple_transparent"
-                        />
-
-                    <include layout="@layout/primary_dark_divider"
-                        android:layout_width="match_parent" android:layout_height="3dp" />
-                </LinearLayout>
-
-            </FrameLayout>
-
-        </LinearLayout>
-        <include android:id="@+id/ad_area" layout="@layout/ad_banner" />
-    </FrameLayout>
-</layout>
