+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -20,11 +20,6 @@
     <item android:title="@string/title_section_settings_this_app">
         <menu>
             <item
-                android:id="@+id/nav_settings"
-                android:icon="@drawable/ic_settings_black"
-                android:title="@string/title_settings"
-                />
-            <item
                 android:id="@+id/nav_gallery"
                 android:icon="@drawable/ic_menu_gallery"
                 android:title="@string/title_set_bg"
@@ -44,6 +39,11 @@
                 android:icon="@drawable/ic_license"
                 android:title=""
                 app:actionLayout="@layout/radio_locale"
+                />
+            <item
+                android:id="@+id/nav_clear_settings"
+                android:icon="@drawable/ic_close_black"
+                android:title="@string/title_clear"
                 />
         </menu>
     </item>
